feat(web): implement creating message, but without getting user id from JWT

+++ src/Tradify.Backend/Tradify.Chat/Tradify.Chat.Domain/Entities/ChatUser.cs
@@ -2,8 +2,8 @@ namespace Tradify.Chat.Domain.Entities;
 
 public class ChatUser
 {
-    public int ChatId { get; set; }
-    public int UserId { get; set; }
+    public long ChatId { get; set; }
+    public long UserId { get; set; }
     
     public Chat Chat { get; set; }
 }
